requirements.txt
matplotlib
names
numpy
pillow
pyyaml
tensorflow
